+++ SPYROpenGL/src/Montage.cpp
@@ -34,46 +34,46 @@
 const void Montage::MontageSpyro(float rotationAiles, float rotationBouche)
 {
     /* param Corps */
-	const float rayonCorps = 0.2;
+	const float rayonCorps = 0.2f;
     const float longueurCorps = 1;
-    const float largeurDUnCoteDuCorps = sqrt(rayonCorps*rayonCorps + rayonCorps*rayonCorps - 2*rayonCorps*rayonCorps*cos(2*M_PI/5));  //théorème d'Al-Kashi : BC²=AB²+AC²-2.AB.AC.cos(BÂC)
+    const float largeurDUnCoteDuCorps = (float) (sqrt(rayonCorps*rayonCorps + rayonCorps*rayonCorps - 2*rayonCorps*rayonCorps*cos(2*M_PI/5)));  //théorème d'Al-Kashi : BC²=AB²+AC²-2.AB.AC.cos(BÂC)
 
     /* param LiaisonCorpsCou */
-    const float coeffLiaisonCorpsCouVersLeCou = 0.5;
-    const float hauteurLiaisonCorpsCou = 0.2;
+    const float coeffLiaisonCorpsCouVersLeCou = 0.5f;
+    const float hauteurLiaisonCorpsCou = 0.2f;
 
     /* param Jambes */
-    float taille = 0.1;
-	float hauteurJambes = 0.2;
+    float taille = 0.1f;
+	float hauteurJambes = 0.2f;
 
 	/* param Ailes */
-	float ecart = 0.05;
-	float largeur = 0.2;
-	float longueur = 0.3;
-	float hauteur = 0.02;
-	float decalageCentre = 0.2;
+	float ecart = 0.05f;
+	float largeur = 0.2f;
+	float longueur = 0.3f;
+	float hauteur = 0.02f;
+	float decalageCentre = 0.2f;
 
 	/* param Cou */
-	float hauteurCou = 0.25;
-	float decalageCou = 0;
+	float hauteurCou = 0.25f;
+	float decalageCou = 0.0f;
 
 	/* param Cornes */
-	float hauteurCorne = 1;
-	float largeurCorne = 0.15;
+	float hauteurCorne = 1.0f;
+	float largeurCorne = 0.15f;
 
 	/* param Tête */
     float tailleTete = 0.5;
 
     /* param Bouche */
-    const float boucheCoeffX = 3;
-    const float boucheCoeffZ = 5;
-    const float boucheInclinaison = -0.1;
+    const float boucheCoeffX = 3.0f;
+    const float boucheCoeffZ = 5.0f;
+    const float boucheInclinaison = -0.1f;
 
     /* param Nez */
-    const float nezCoeffIY = 8;
+    const float nezCoeffIY = 8.0f;
     glPushMatrix();
         glTranslatef(0,taille+2*hauteurJambes,0); // translation verticale de la hauteur de Jambe+Pied et translation horizontale de la largeur de la jambe
-        glTranslatef(0,0,-3*sqrt(pow((sqrt((taille*taille)/2)),2)-pow(taille/2,2))); // translation pour placer le corps en adéquation avec les pieds
+        glTranslatef(0,0,(float) ( - 3 * sqrt(pow((sqrt((taille * taille) / 2)), 2) - pow(taille / 2, 2)))); // translation pour placer le corps en adéquation avec les pieds
 
         /* Création du corps */
         std::tuple<Point, Point> deuxPointsBaseCorps = Creation::Corps(5,rayonCorps,longueurCorps);
@@ -92,17 +92,17 @@ const void Montage::MontageSpyro(float rotationAiles, float rotationBouche)
             //Bouche (et menton)
             //Déplacement de l'origine sur le milieu de la surface haute du cou
             glTranslatef(0, hauteurCou, 0);
-            Dessin::Bouche(rayonCorps*coeffLiaisonCorpsCouVersLeCou, tailleTete*0.3, boucheCoeffX, boucheCoeffZ, tailleTete*0.2, tailleTete*boucheInclinaison);
+            Dessin::Bouche(rayonCorps*coeffLiaisonCorpsCouVersLeCou, tailleTete*0.3f, boucheCoeffX, boucheCoeffZ, tailleTete*0.2f, tailleTete*boucheInclinaison);
 
             //Nez
             //Déplacement de l'origine sur le "milieu" (il n'y a pas réellement de point à cet endroit) du prolongement du cou utilisé pour la surface haute de la bouche
-            glTranslatef(0,tailleTete*0.3,tailleTete*0.2);
+            glTranslatef(0,tailleTete*0.3f,tailleTete*0.2f);
             glRotatef(rotationBouche,-1,0,0); // animation bouche Spyro (quand appui sur ESPACE)
-            Dessin::Nez(rayonCorps*coeffLiaisonCorpsCouVersLeCou, tailleTete*0.1, boucheCoeffX, boucheCoeffZ, nezCoeffIY, 0);
+            Dessin::Nez(rayonCorps*coeffLiaisonCorpsCouVersLeCou, tailleTete*0.1f, boucheCoeffX, boucheCoeffZ, nezCoeffIY, 0);
         glPopMatrix();
         /* Création ailes */
         glPushMatrix();
-            glTranslatef(0,0,longueurCorps*0.7);
+            glTranslatef(0,0,longueurCorps*0.7f);
             Creation::Aile(ecart, largeur, longueur, hauteur, decalageCentre, rotationAiles);
         glPopMatrix();
     glPopMatrix();
